split up /lua/spreadattack.lua into sim / ui parts
